Update: path 'solution/solution_1.cc' - Updated by 'nNaxhO'

+++ solution/solution_1.cc
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-const double PI = acos(-1);
+const float PI = acos(-1);
 
 void FFT(vector<complex<float>>& X) {
     int n = X.size();
